test(shim): dedup capture-server helper, harden recv_timeout, document clone-not-take (#369)
- Extract the duplicated spawn_capture_server into tests/common/mod.rs.
- Fail loudly on RecvTimeoutError::Disconnected (capture-harness death) instead
  of treating it as "no telemetry" — only Timeout means no telemetry.
- Document why force_flush_bounded clones (not takes) the providers: Drop is the
  shutdown gate and must retain them.
