+++ tests/shim_bounded_drop_dataloss_pentest.rs
@@ -16,41 +16,14 @@
 //! an OTLP export must arrive well within the test window. If nothing arrives, the
 //! bounded Drop is dropping telemetry in the common case — a regression.
 
-use std::io::{Read, Write};
-use std::net::TcpListener;
+mod common;
+
 use std::os::unix::fs::symlink;
 use std::process::Command;
-use std::sync::mpsc;
+use std::sync::mpsc::RecvTimeoutError;
 use std::time::{Duration, Instant};
 
 use tempfile::TempDir;
-
-/// One-shot HTTP capture server that answers immediately (healthy collector).
-/// Returns the base URL and a receiver that yields `true` on the first request body.
-fn spawn_capture_server() -> (String, mpsc::Receiver<bool>) {
-    let listener = TcpListener::bind("127.0.0.1:0").unwrap();
-    let addr = listener.local_addr().unwrap();
-    let base = format!("http://{}", addr);
-    let (tx, rx) = mpsc::channel();
-
-    std::thread::spawn(move || {
-        for stream in listener.incoming() {
-            let Ok(mut stream) = stream else { continue };
-            stream
-                .set_read_timeout(Some(Duration::from_millis(500)))
-                .ok();
-            let mut buf = [0u8; 4096];
-            let got = matches!(stream.read(&mut buf), Ok(n) if n > 0);
-            let _ = stream
-                .write_all(b"HTTP/1.1 200 OK\r\nContent-Length: 0\r\nConnection: close\r\n\r\n");
-            if got {
-                let _ = tx.send(true);
-            }
-        }
-    });
-
-    (base, rx)
-}
 
 fn git(repo: &std::path::Path, args: &[&str]) {
     let out = Command::new("git")
@@ -94,7 +67,7 @@ fn structured_diff_still_exports_to_a_healthy_collector_under_bounded_teardown()
     let real_path = std::env::var("PATH").unwrap_or_default();
     let path = format!("{}:{}", shim_dir.display(), real_path);
 
-    let (endpoint, rx) = spawn_capture_server();
+    let (endpoint, rx) = common::spawn_capture_server();
 
     let start = Instant::now();
     let out = Command::new(&shim_git)
@@ -116,7 +89,18 @@ fn structured_diff_still_exports_to_a_healthy_collector_under_bounded_teardown()
     );
 
     // The bounded teardown must still deliver the export to a healthy collector.
-    let received = rx.recv_timeout(Duration::from_secs(5)).unwrap_or(false);
+    //
+    // RecvTimeoutError::Timeout  → no export arrived within the window (test fails below)
+    // RecvTimeoutError::Disconnected → capture-server thread died; this is a harness
+    //   failure, NOT a "no telemetry" signal — panic loudly so the failure is not
+    //   silently misread as a data-loss regression in the bounded teardown.
+    let received = match rx.recv_timeout(Duration::from_secs(5)) {
+        Ok(val) => val,
+        Err(RecvTimeoutError::Timeout) => false,
+        Err(RecvTimeoutError::Disconnected) => {
+            panic!("capture-server thread disconnected — harness failure, not a telemetry signal")
+        }
+    };
     assert!(
         received,
         "bounded Drop (500ms) lost telemetry against a HEALTHY collector: no OTLP \
