+++ tests/shim_passthrough_optout_pentest.rs
@@ -11,41 +11,14 @@
 //!
 //! These drive the built binary end-to-end through argv[0]="git" shim mode.
 
-use std::io::{Read, Write};
-use std::net::TcpListener;
+mod common;
+
 use std::os::unix::fs::{PermissionsExt, symlink};
 use std::process::Command;
-use std::sync::mpsc;
+use std::sync::mpsc::RecvTimeoutError;
 use std::time::Duration;
 
 use tempfile::TempDir;
-
-/// Spawn a one-shot HTTP capture server. Returns the bound `base` URL and a
-/// receiver that yields `true` once ANY HTTP request body is received.
-fn spawn_capture_server() -> (String, mpsc::Receiver<bool>) {
-    let listener = TcpListener::bind("127.0.0.1:0").unwrap();
-    let addr = listener.local_addr().unwrap();
-    let base = format!("http://{}", addr);
-    let (tx, rx) = mpsc::channel();
-
-    std::thread::spawn(move || {
-        for stream in listener.incoming() {
-            let Ok(mut stream) = stream else { continue };
-            stream
-                .set_read_timeout(Some(Duration::from_millis(500)))
-                .ok();
-            let mut buf = [0u8; 4096];
-            let got = matches!(stream.read(&mut buf), Ok(n) if n > 0);
-            let _ = stream
-                .write_all(b"HTTP/1.1 200 OK\r\nContent-Length: 0\r\nConnection: close\r\n\r\n");
-            if got {
-                let _ = tx.send(true);
-            }
-        }
-    });
-
-    (base, rx)
-}
 
 /// Create a fake real `git` that prints a unique marker and exits 0. Lets us
 /// prove the REAL git ran (not the shim) and that output is byte-identical.
@@ -131,7 +104,7 @@ fn it_emits_no_otlp_export_when_passthrough_opt_out_is_set() {
     make_marker_git(&real_dir, "X");
 
     let path = format!("{}:{}", shim_dir.display(), real_dir.display());
-    let (endpoint, rx) = spawn_capture_server();
+    let (endpoint, rx) = common::spawn_capture_server();
 
     let status = Command::new(&shim_git)
         .env("GIT_PRISM_PASSTHROUGH", "1")
@@ -147,7 +120,18 @@ fn it_emits_no_otlp_export_when_passthrough_opt_out_is_set() {
 
     // A correct opt-out never initializes telemetry, so NO export should arrive.
     // If one arrives within 2 s, telemetry leaked on the opted-out path.
-    let leaked = rx.recv_timeout(Duration::from_secs(2)).unwrap_or(false);
+    //
+    // RecvTimeoutError::Timeout  → no telemetry within the window (expected)
+    // RecvTimeoutError::Disconnected → capture-server thread died; this is a
+    //   harness failure, NOT a "no telemetry" signal — panic loudly so the
+    //   failure isn't silently misread as a passing opt-out test.
+    let leaked = match rx.recv_timeout(Duration::from_secs(2)) {
+        Ok(val) => val,
+        Err(RecvTimeoutError::Timeout) => false,
+        Err(RecvTimeoutError::Disconnected) => {
+            panic!("capture-server thread disconnected — harness failure, not a telemetry signal")
+        }
+    };
     assert!(
         !leaked,
         "opted-out invocation emitted an OTLP export; telemetry must be skipped \
